+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 
 # 更新
 
-## 0.3.0-beta.1 (none)
+## 0.3.0 (none)
 
-* update ng-alain 1.0.0-beta.4
-* update ng-zorro-antd 0.7.0-beta.2
+* update ng-alain 1.0.4
+* update ng-zorro-antd 0.7.1
 * ADD: minitable isEditCell
 * FIX: minitable editCell
 * 重构：sip-card
 * ADD: minitable tree mode
+* update: sip-helper 0.1.0
 
 ## 0.2.5